requirements.txt
pillow
pytesseract
pandas
flask
streamlit